requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.dev.in
#
astroid==3.0.2
    # via pylint
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
coverage[toml]==6.5.0
    # via
    #   coverage
    #   coveralls
    #   pytest-cov
coveralls==3.3.1
    # via -r requirements.dev.in
dill==0.3.7
    # via pylint
distlib==0.3.8
    # via virtualenv
docopt==0.6.2
    # via coveralls
faker==21.0.0
    # via -r requirements.dev.in
filelock==3.13.1
    # via virtualenv
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pylint
mccabe==0.7.0
    # via pylint
mypy==1.7.1
    # via -r requirements.dev.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.8.0
    # via pre-commit
packaging==23.2
    # via
    #   build
    #   pytest
pip-tools==7.3.0
    # via -r requirements.dev.in
platformdirs==4.1.0
    # via
    #   pylint
    #   virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.dev.in
pylint==3.0.3
    # via
    #   -r requirements.dev.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r requirements.dev.in
pylint-plugin-utils==0.8.2
    # via pylint-django
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via
    #   -r requirements.dev.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements.dev.in
pytest-django==4.7.0
    # via -r requirements.dev.in
python-dateutil==2.8.2
    # via faker
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via coveralls
six==1.16.0
    # via python-dateutil
tomlkit==0.12.3
    # via pylint
typing-extensions==4.9.0
    # via mypy
urllib3==2.1.0
    # via requests
virtualenv==20.25.0
    # via pre-commit
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
